+++ src/aws_organization/lib/__init__.py
@@ -1,5 +1,7 @@
 from .account import AwsAccount
+from .org_units import OrganizationalUnits
 from .org_units import create_organizational_units
 from .workload import DEFAULT_ORG_ACCESS_ROLE_NAME
 from .workload import AwsWorkload
 from .workload import CommonWorkloadKwargs
+from .workload import create_pulumi_kms_role_policy_args
