+++ src/aws_organization/program.py
@@ -31,11 +31,12 @@ from .lib import create_organizational_units
 from .lib.shared_lib import WORKLOAD_INFO_SSM_PARAM_PREFIX
 from .lib.shared_lib import AwsAccountInfo
 from .lib.shared_lib import AwsLogicalWorkload
+from .workloads import create_workloads
 
 logger = logging.getLogger(__name__)
 
 
-def pulumi_program() -> None:  # noqa: PLR0915 # yes, this is getting long...need to refactor soon
+def pulumi_program() -> None:
     """Execute creating the stack."""
     aws_account_id = get_aws_account_id()
     export("aws-account-id", aws_account_id)
@@ -291,82 +292,6 @@ def pulumi_program() -> None:  # noqa: PLR0915 # yes, this is getting long...nee
         )
     )
 
-    biotasker_dev_account = AwsAccount(ou=org_units.non_qualified_workload_dev, account_name="biotasker-dev")
-
-    dev_account_data = [biotasker_dev_account.account_info_kwargs]
-    all_dev_accounts_resolved = Output.all(
-        *[
-            Output.all(account["id"], account["name"]).apply(lambda vals: {"id": vals[0], "name": vals[1]})
-            for account in dev_account_data
-        ]
-    )
-
-    def build_workload(resolved_accounts: list[dict[str, str]]) -> str:
-        # Convert resolved dicts to Pydantic AwsAccountInfo models
-        dev_accounts = [AwsAccountInfo(**acc) for acc in resolved_accounts]
-
-        logical_workload = AwsLogicalWorkload(
-            name="biotasker",
-            dev_accounts=dev_accounts,  # Insert all resolved dev accounts
-        )
-
-        return logical_workload.model_dump_json()
-
-    workload_name = "biotasker"
-    _ = ssm.Parameter(  # TODO: consider DRY-ing this up with the parameter generation in lib.py
-        f"{workload_name}-workload-info-for-central-infra",
-        description="Hold the logical workload information so that Central Infra account can deploy various resources within them.",
-        type=ssm.ParameterType.STRING,
-        name=f"{WORKLOAD_INFO_SSM_PARAM_PREFIX}/{workload_name}",
-        tags=common_tags(),
-        value=all_dev_accounts_resolved.apply(build_workload),
-        opts=ResourceOptions(provider=central_infra_provider, parent=central_infra_account),
-    )
-    biotasker_role_arn = biotasker_dev_account.account.id.apply(
-        lambda x: f"arn:aws:iam::{x}:role/{DEFAULT_ORG_ACCESS_ROLE_NAME}"
-    )
-    assume_role = ProviderAssumeRoleArgs(role_arn=biotasker_role_arn, session_name="blah")
-    biotasker_provider = Provider(
-        "biotasker-dev",
-        assume_role=assume_role,
-        region="us-east-1",
-        opts=ResourceOptions(
-            parent=biotasker_dev_account,
-        ),
-    )
-    _ = iam.Role(
-        f"central-infra-repo-deploy-in-{workload_name}",
-        role_name=f"InfraDeploy--{CENTRAL_INFRA_REPO_NAME}",
-        assume_role_policy_document=deploy_in_workload_account_assume_role_policy.json,
-        managed_policy_arns=["arn:aws:iam::aws:policy/AdministratorAccess"],
-        tags=common_tags_native(),
-        opts=ResourceOptions(provider=biotasker_provider, parent=biotasker_dev_account),
-    )
-    _ = iam.Role(
-        f"central-infra-repo-preview-in-{workload_name}",
-        role_name=f"InfraPreview--{CENTRAL_INFRA_REPO_NAME}",
-        assume_role_policy_document=preview_in_workload_account_assume_role_policy.json,
-        managed_policy_arns=["arn:aws:iam::aws:policy/ReadOnlyAccess"],
-        policies=[
-            iam.RolePolicyArgs(
-                policy_name="InfraKmsDecrypt",
-                policy_document=get_policy_document(
-                    statements=[
-                        GetPolicyDocumentStatementArgs(
-                            effect="Allow",
-                            actions=[
-                                "kms:Decrypt",
-                                "kms:Encrypt",  # unclear why Encrypt is required to run a Preview...but Pulumi gives an error if it's not included
-                            ],
-                            resources=[kms_key_arn],
-                        )
-                    ]
-                ).json,
-            )
-        ],
-        tags=common_tags_native(),
-        opts=ResourceOptions(provider=biotasker_provider, parent=biotasker_dev_account),
-    )
     common_workload_kwargs: CommonWorkloadKwargs = {
         "central_infra_account": central_infra_account,
         "deploy_in_workload_account_assume_role_policy": deploy_in_workload_account_assume_role_policy,
@@ -394,6 +319,7 @@ def pulumi_program() -> None:  # noqa: PLR0915 # yes, this is getting long...nee
             ],
         ),
     )
+    create_workloads(org_units=org_units, common_workload_kwargs=common_workload_kwargs)
     if CONFIGURE_CLOUD_COURIER:
         _ = AwsWorkload(
             workload_name="cloud-courier",
